@@ -4,31 +4,30 @@ using Meadow.Peripherals.Displays;
 using Meadow.Peripherals.Leds;
 using Meadow.Peripherals.Sensors.Light;
 
-namespace AmbientRoomMonitor.Hardware
+namespace AmbientRoomMonitor.Hardware;
+
+internal class AmbientRoomMonitorHardware : IAmbientRoomMonitorHardware
 {
-    internal class AmbientRoomMonitorHardware : IAmbientRoomMonitorHardware
+    protected IProjectLabHardware ProjLab { get; private set; }
+
+    public IPixelDisplay Display { get; set; }
+
+    public IRgbPwmLed RgbPwmLed { get; set; }
+
+    public ILightSensor LightSensor { get; set; }
+
+    public Bme68x EnvironmentalSensor { get; set; }
+
+    public void Initialize()
     {
-        protected IProjectLabHardware ProjLab { get; private set; }
+        ProjLab = ProjectLab.Create();
 
-        public IPixelDisplay Display { get; set; }
+        Display = ProjLab.Display;
 
-        public IRgbPwmLed RgbPwmLed { get; set; }
+        RgbPwmLed = ProjLab.RgbLed;
 
-        public ILightSensor LightSensor { get; set; }
+        LightSensor = ProjLab.LightSensor;
 
-        public Bme68x EnvironmentalSensor { get; set; }
-
-        public void Initialize()
-        {
-            ProjLab = ProjectLab.Create();
-
-            Display = ProjLab.Display;
-
-            RgbPwmLed = ProjLab.RgbLed;
-
-            LightSensor = ProjLab.LightSensor;
-
-            EnvironmentalSensor = (ProjLab as ProjectLabHardwareBase).AtmosphericSensor;
-        }
+        EnvironmentalSensor = (ProjLab as ProjectLabHardwareBase).AtmosphericSensor;
     }
 }
