@@ -3,45 +3,44 @@ using AmbientRoomMonitor.Services;
 using Meadow;
 using System;
 
-namespace AmbientRoomMonitor
+namespace AmbientRoomMonitor;
+
+internal class MainController
 {
-    internal class MainController
+    IAmbientRoomMonitorHardware hardware;
+
+    DisplayController displayService;
+
+    public MainController(IAmbientRoomMonitorHardware hardware)
     {
-        IAmbientRoomMonitorHardware hardware;
+        this.hardware = hardware;
+    }
 
-        DisplayController displayService;
+    public void Initialize()
+    {
+        hardware.Initialize();
 
-        public MainController(IAmbientRoomMonitorHardware hardware)
-        {
-            this.hardware = hardware;
-        }
+        displayService = new DisplayController(hardware.Display);
 
-        public void Initialize()
-        {
-            hardware.Initialize();
+        hardware.EnvironmentalSensor.Updated += EnvironmentalSensorUpdated;
+    }
 
-            displayService = new DisplayController(hardware.Display);
+    private void EnvironmentalSensorUpdated(object sender, Meadow.IChangeResult<(Meadow.Units.Temperature? Temperature, Meadow.Units.RelativeHumidity? Humidity, Meadow.Units.Pressure? Pressure, Meadow.Units.Resistance? GasResistance)> e)
+    {
+        hardware.RgbPwmLed.StartBlink(Color.Orange);
 
-            hardware.EnvironmentalSensor.Updated += EnvironmentalSensorUpdated;
-        }
+        displayService.UpdateAtmosphericConditions(
+            light: $"{hardware.LightSensor.Illuminance.Value.Lux:N0}",
+            pressure: $"{e.New.Pressure.Value.Millibar:N0}",
+            humidity: $"{e.New.Humidity.Value.Percent:N0}",
+            temperature: $"{e.New.Temperature.Value.Celsius:N0}");
 
-        private void EnvironmentalSensorUpdated(object sender, Meadow.IChangeResult<(Meadow.Units.Temperature? Temperature, Meadow.Units.RelativeHumidity? Humidity, Meadow.Units.Pressure? Pressure, Meadow.Units.Resistance? GasResistance)> e)
-        {
-            hardware.RgbPwmLed.StartBlink(Color.Orange);
+        hardware.RgbPwmLed.StartBlink(Color.Green);
+    }
 
-            displayService.UpdateAtmosphericConditions(
-                light: $"{hardware.LightSensor.Illuminance.Value.Lux:N0}",
-                pressure: $"{e.New.Pressure.Value.Millibar:N0}",
-                humidity: $"{e.New.Humidity.Value.Percent:N0}",
-                temperature: $"{e.New.Temperature.Value.Celsius:N0}");
-
-            hardware.RgbPwmLed.StartBlink(Color.Green);
-        }
-
-        public void Run()
-        {
-            hardware.LightSensor.StartUpdating(TimeSpan.FromSeconds(5));
-            hardware.EnvironmentalSensor.StartUpdating(TimeSpan.FromSeconds(5));
-        }
+    public void Run()
+    {
+        hardware.LightSensor.StartUpdating(TimeSpan.FromSeconds(5));
+        hardware.EnvironmentalSensor.StartUpdating(TimeSpan.FromSeconds(5));
     }
 }
