@@ -2,16 +2,15 @@
 using Meadow.Peripherals.Displays;
 using Meadow.Peripherals.Leds;
 
-namespace MagicEightMeadow.Hardware
+namespace MagicEightMeadow.Hardware;
+
+internal interface IMagicEightMeadowHardware
 {
-    internal interface IMagicEightMeadowHardware
-    {
-        public IPixelDisplay Display { get; }
+    public IPixelDisplay Display { get; }
 
-        public IRgbPwmLed RgbPwmLed { get; }
+    public IRgbPwmLed RgbPwmLed { get; }
 
-        public Bmi270 MotionSensor { get; }
+    public Bmi270 MotionSensor { get; }
 
-        public void Initialize();
-    }
+    public void Initialize();
 }
