@@ -3,27 +3,26 @@ using Meadow.Foundation.Sensors.Motion;
 using Meadow.Peripherals.Displays;
 using Meadow.Peripherals.Leds;
 
-namespace MagicEightMeadow.Hardware
+namespace MagicEightMeadow.Hardware;
+
+internal class MagicEightMeadowHardware : IMagicEightMeadowHardware
 {
-    internal class MagicEightMeadowHardware : IMagicEightMeadowHardware
+    protected IProjectLabHardware projectLab { get; private set; }
+
+    public IPixelDisplay Display { get; set; }
+
+    public IRgbPwmLed RgbPwmLed { get; set; }
+
+    public Bmi270 MotionSensor { get; set; }
+
+    public void Initialize()
     {
-        protected IProjectLabHardware projectLab { get; private set; }
+        projectLab = ProjectLab.Create();
 
-        public IPixelDisplay Display { get; set; }
+        Display = projectLab.Display;
 
-        public IRgbPwmLed RgbPwmLed { get; set; }
+        RgbPwmLed = projectLab.RgbLed;
 
-        public Bmi270 MotionSensor { get; set; }
-
-        public void Initialize()
-        {
-            projectLab = ProjectLab.Create();
-
-            Display = projectLab.Display;
-
-            RgbPwmLed = projectLab.RgbLed;
-
-            MotionSensor = (projectLab as ProjectLabHardwareBase).MotionSensor;
-        }
+        MotionSensor = (projectLab as ProjectLabHardwareBase).MotionSensor;
     }
 }
