@@ -6,60 +6,59 @@ using Meadow.Units;
 using System;
 using System.Threading.Tasks;
 
-namespace MagicEightMeadow
+namespace MagicEightMeadow;
+
+internal class MainController
 {
-    internal class MainController
+    private bool isAnswering;
+    private IMagicEightMeadowHardware hardware;
+    private DisplayController displayController;
+
+    public MainController(IMagicEightMeadowHardware hardware)
     {
-        private bool isAnswering;
-        private IMagicEightMeadowHardware hardware;
-        private DisplayController displayController;
+        this.hardware = hardware;
+    }
 
-        public MainController(IMagicEightMeadowHardware hardware)
-        {
-            this.hardware = hardware;
-        }
+    public void Initialize()
+    {
+        hardware.Initialize();
 
-        public void Initialize()
-        {
-            hardware.Initialize();
+        displayController = new DisplayController(hardware.Display);
 
-            displayController = new DisplayController(hardware.Display);
+        var consumer = Bmi270.CreateObserver(
+            handler: result => MotionSensorHandler(result),
+            filter: result => MotionSensorFilter(result));
+        hardware.MotionSensor.Subscribe(consumer);
+    }
 
-            var consumer = Bmi270.CreateObserver(
-                handler: result => MotionSensorHandler(result),
-                filter: result => MotionSensorFilter(result));
-            hardware.MotionSensor.Subscribe(consumer);
-        }
+    private async void MotionSensorHandler(IChangeResult<(Acceleration3D? a3D, AngularVelocity3D? v3D, Temperature? t)> e)
+    {
+        if (isAnswering)
+            return;
+        isAnswering = true;
 
-        private async void MotionSensorHandler(IChangeResult<(Acceleration3D? a3D, AngularVelocity3D? v3D, Temperature? t)> e)
-        {
-            if (isAnswering)
-                return;
-            isAnswering = true;
+        hardware.RgbPwmLed.SetColor(Color.Orange);
 
-            hardware.RgbPwmLed.SetColor(Color.Orange);
+        displayController.ShowAnswer();
 
-            displayController.ShowAnswer();
+        await Task.Delay(TimeSpan.FromSeconds(5));
 
-            await Task.Delay(TimeSpan.FromSeconds(5));
+        displayController.ShowQuestion();
 
-            displayController.ShowQuestion();
+        hardware.RgbPwmLed.SetColor(Color.Green);
 
-            hardware.RgbPwmLed.SetColor(Color.Green);
+        isAnswering = false;
+    }
 
-            isAnswering = false;
-        }
+    private bool MotionSensorFilter(IChangeResult<(Acceleration3D? a3D, AngularVelocity3D? v3D, Temperature? t)> e)
+    {
+        return e.New.v3D.Value.X.DegreesPerSecond > 0.75 || e.New.v3D.Value.Y.DegreesPerSecond > 0.75 || e.New.v3D.Value.Z.DegreesPerSecond > 0.75;
+    }
 
-        private bool MotionSensorFilter(IChangeResult<(Acceleration3D? a3D, AngularVelocity3D? v3D, Temperature? t)> e)
-        {
-            return e.New.v3D.Value.X.DegreesPerSecond > 0.75 || e.New.v3D.Value.Y.DegreesPerSecond > 0.75 || e.New.v3D.Value.Z.DegreesPerSecond > 0.75;
-        }
+    public void Run()
+    {
+        displayController.ShowQuestion();
 
-        public void Run()
-        {
-            displayController.ShowQuestion();
-
-            hardware.MotionSensor.StartUpdating(TimeSpan.FromSeconds(1));
-        }
+        hardware.MotionSensor.StartUpdating(TimeSpan.FromSeconds(1));
     }
 }
