@@ -5,94 +5,93 @@ using Meadow.Peripherals.Displays;
 using System;
 using System.Threading.Tasks;
 
-namespace MathLineCharts
+namespace MathLineCharts;
+
+public class MeadowApp : App<F7CoreComputeV2>
 {
-    public class MeadowApp : App<F7CoreComputeV2>
+    private DisplayScreen _screen;
+
+    public override Task Initialize()
     {
-        private DisplayScreen _screen;
+        Resolver.Log.Info("Initialize...");
 
-        public override Task Initialize()
+        var projectLab = ProjectLab.Create();
+
+        _screen = new DisplayScreen(projectLab.Display, RotationType._270Degrees);
+
+        _screen.BackgroundColor = Color.AntiqueWhite;
+
+        var chart1 = new LineChart(0, 0, _screen.Width, _screen.Height / 2)
         {
-            Resolver.Log.Info("Initialize...");
+            BackgroundColor = Color.FromHex("111111"),
+            ShowYAxisLabels = true
+        };
 
-            var projectLab = ProjectLab.Create();
+        chart1.Series.Add(
+            GetSineSeries(),
+            GetCosineSeries(4, 4.2, 0));
 
-            _screen = new DisplayScreen(projectLab.Display, RotationType._270Degrees);
+        var chart2 = new LineChart(0, _screen.Height / 2, _screen.Width, _screen.Height / 2)
+        {
+            BackgroundColor = Color.FromHex("222222"),
+            ShowYAxisLabels = true
+        };
 
-            _screen.BackgroundColor = Color.AntiqueWhite;
+        chart2.Series.Add(
+            GetSineSeries(2, 2),
+            GetCosineSeries(4, 4.2, 4.5));
 
-            var chart1 = new LineChart(0, 0, _screen.Width, _screen.Height / 2)
-            {
-                BackgroundColor = Color.FromHex("111111"),
-                ShowYAxisLabels = true
-            };
+        _screen.Controls.Add(chart1, chart2);
 
-            chart1.Series.Add(
-                GetSineSeries(),
-                GetCosineSeries(4, 4.2, 0));
+        return base.Initialize();
+    }
 
-            var chart2 = new LineChart(0, _screen.Height / 2, _screen.Width, _screen.Height / 2)
-            {
-                BackgroundColor = Color.FromHex("222222"),
-                ShowYAxisLabels = true
-            };
+    private const int PointsPerSeries = 50;
 
-            chart2.Series.Add(
-                GetSineSeries(2, 2),
-                GetCosineSeries(4, 4.2, 4.5));
+    private LineChartSeries GetSineSeries(double xScale = 4, double yScale = 1.5, double yOffset = 1.5)
+    {
+        var series = new LineChartSeries()
+        {
+            LineColor = Color.Red,
+            PointColor = Color.Green,
+            LineStroke = 1,
+            PointSize = 2,
+            ShowLines = true,
+            ShowPoints = true,
+        };
 
-            _screen.Controls.Add(chart1, chart2);
-
-            return base.Initialize();
+        for (var p = 0; p < PointsPerSeries; p++)
+        {
+            series.Points.Add(p * 2, (Math.Sin(p / xScale) * yScale) + yOffset);
         }
 
-        private const int PointsPerSeries = 50;
+        return series;
+    }
 
-        private LineChartSeries GetSineSeries(double xScale = 4, double yScale = 1.5, double yOffset = 1.5)
+    private LineChartSeries GetCosineSeries(double xScale = 4, double yScale = 1.5, double yOffset = 4.5)
+    {
+        var series = new LineChartSeries()
         {
-            var series = new LineChartSeries()
-            {
-                LineColor = Color.Red,
-                PointColor = Color.Green,
-                LineStroke = 1,
-                PointSize = 2,
-                ShowLines = true,
-                ShowPoints = true,
-            };
+            LineColor = Color.DarkBlue,
+            PointColor = Color.DarkGreen,
+            LineStroke = 2,
+            PointSize = 3,
+            ShowLines = true,
+            ShowPoints = true,
+        };
 
-            for (var p = 0; p < PointsPerSeries; p++)
-            {
-                series.Points.Add(p * 2, (Math.Sin(p / xScale) * yScale) + yOffset);
-            }
-
-            return series;
+        for (var p = 0; p < PointsPerSeries; p++)
+        {
+            series.Points.Add(p * 2, (Math.Cos(p / xScale) * yScale) + yOffset);
         }
 
-        private LineChartSeries GetCosineSeries(double xScale = 4, double yScale = 1.5, double yOffset = 4.5)
-        {
-            var series = new LineChartSeries()
-            {
-                LineColor = Color.DarkBlue,
-                PointColor = Color.DarkGreen,
-                LineStroke = 2,
-                PointSize = 3,
-                ShowLines = true,
-                ShowPoints = true,
-            };
+        return series;
+    }
 
-            for (var p = 0; p < PointsPerSeries; p++)
-            {
-                series.Points.Add(p * 2, (Math.Cos(p / xScale) * yScale) + yOffset);
-            }
+    public override Task Run()
+    {
+        Resolver.Log.Info("Run...");
 
-            return series;
-        }
-
-        public override Task Run()
-        {
-            Resolver.Log.Info("Run...");
-
-            return base.Run();
-        }
+        return base.Run();
     }
 }
