@@ -1,48 +1,45 @@
 ﻿using Meadow;
 using Meadow.Devices;
-using Meadow.Foundation;
 using Meadow.Foundation.Grove.Sensors.Moisture;
-using Meadow.Foundation.Leds;
 using Meadow.Peripherals.Leds;
 using MoistureMeter.Controllers;
 using System;
 using System.Threading.Tasks;
 
-namespace MoistureMeter
+namespace MoistureMeter;
+
+// Change F7CoreComputeV2 to F7FeatherV2 for ProjectLab v2
+public class MeadowApp : App<F7CoreComputeV2>
 {
-    // Change F7CoreComputeV2 to F7FeatherV2 for ProjectLab v2
-    public class MeadowApp : App<F7CoreComputeV2>
+    private IRgbPwmLed onboardLed;
+    private MoistureSensor sensor;
+    private IProjectLabHardware projectLab;
+
+    public override Task Initialize()
     {
-        private IRgbPwmLed onboardLed;
-        private MoistureSensor sensor;
-        private IProjectLabHardware projectLab;
+        Resolver.Log.Info("Initialize...");
 
-        public override Task Initialize()
+        projectLab = ProjectLab.Create();
+        Resolver.Log.Info($"Running on ProjectLab Hardware {projectLab.RevisionString}");
+
+        onboardLed = projectLab.RgbLed;
+        onboardLed.SetColor(Color.Red);
+
+        DisplayController.Instance.Initialize(projectLab.Display);
+
+        sensor = new MoistureSensor(projectLab.GroveAnalog.Pins.D0);
+
+        sensor.Updated += (sender, result) =>
         {
-            Resolver.Log.Info("Initialize...");
+            var percentage = (int)ExtensionMethods.Map(result.New.Millivolts, 0, 1750, 0, 100);
 
-            projectLab = ProjectLab.Create();
-            Resolver.Log.Info($"Running on ProjectLab Hardware {projectLab.RevisionString}");
+            DisplayController.Instance.UpdatePercentage(Math.Clamp(percentage, 0, 100));
+        };
 
-            onboardLed = projectLab.RgbLed;
-            onboardLed.SetColor(Color.Red);
+        sensor.StartUpdating(TimeSpan.FromMilliseconds(1000));
 
-            DisplayController.Instance.Initialize(projectLab.Display);
+        onboardLed.SetColor(Color.Green);
 
-            sensor = new MoistureSensor(projectLab.GroveAnalog.Pins.D0);
-
-            sensor.Updated += (sender, result) =>
-            {
-                var percentage = (int)ExtensionMethods.Map(result.New.Millivolts, 0, 1750, 0, 100);
-
-                DisplayController.Instance.UpdatePercentage(Math.Clamp(percentage, 0, 100));
-            };
-
-            sensor.StartUpdating(TimeSpan.FromMilliseconds(1000));
-
-            onboardLed.SetColor(Color.Green);
-
-            return base.Initialize();
-        }
+        return base.Initialize();
     }
 }
