@@ -1,63 +1,62 @@
-﻿namespace WifiWeather.DTOs
+﻿namespace WifiWeather.DTOs;
+
+public class WeatherReadingDTO
 {
-    public class WeatherReadingDTO
-    {
-        public Coordinates coord { get; set; }
-        public Weather[] weather { get; set; }
-        public WeatherValues main { get; set; }
-        public int visibility { get; set; }
-        public Wind wind { get; set; }
-        public Clouds clouds { get; set; }
-        public int dt { get; set; }
-        public System sys { get; set; }
-        public long timezone { get; set; }
-        public int id { get; set; }
-        public string name { get; set; }
-        public int cod { get; set; }
-    }
+    public Coordinates coord { get; set; }
+    public Weather[] weather { get; set; }
+    public WeatherValues main { get; set; }
+    public int visibility { get; set; }
+    public Wind wind { get; set; }
+    public Clouds clouds { get; set; }
+    public int dt { get; set; }
+    public System sys { get; set; }
+    public long timezone { get; set; }
+    public int id { get; set; }
+    public string name { get; set; }
+    public int cod { get; set; }
+}
 
-    public class Coordinates
-    {
-        public double lon { get; set; }
-        public double lat { get; set; }
-    }
+public class Coordinates
+{
+    public double lon { get; set; }
+    public double lat { get; set; }
+}
 
-    public class Weather
-    {
-        public int id { get; set; }
-        public string nain { get; set; }
-        public string description { get; set; }
-        public string icon { get; set; }
-    }
+public class Weather
+{
+    public int id { get; set; }
+    public string nain { get; set; }
+    public string description { get; set; }
+    public string icon { get; set; }
+}
 
-    public class WeatherValues
-    {
-        public double temp { get; set; }
-        public double feels_like { get; set; }
-        public double temp_min { get; set; }
-        public double temp_max { get; set; }
-        public int pressure { get; set; }
-        public int humidity { get; set; }
-    }
+public class WeatherValues
+{
+    public double temp { get; set; }
+    public double feels_like { get; set; }
+    public double temp_min { get; set; }
+    public double temp_max { get; set; }
+    public int pressure { get; set; }
+    public int humidity { get; set; }
+}
 
-    public class Wind
-    {
-        public decimal speed { get; set; }
-        public int deg { get; set; }
-        public double gust { get; set; }
-    }
+public class Wind
+{
+    public decimal speed { get; set; }
+    public int deg { get; set; }
+    public double gust { get; set; }
+}
 
-    public class Clouds
-    {
-        public int all { get; set; }
-    }
+public class Clouds
+{
+    public int all { get; set; }
+}
 
-    public class System
-    {
-        public int type { get; set; }
-        public int id { get; set; }
-        public string country { get; set; }
-        public long sunrise { get; set; }
-        public long sunset { get; set; }
-    }
+public class System
+{
+    public int type { get; set; }
+    public int id { get; set; }
+    public string country { get; set; }
+    public long sunrise { get; set; }
+    public long sunset { get; set; }
 }
