@@ -7,147 +7,146 @@ using System.Threading.Tasks;
 using WifiWeather.Controllers;
 using WifiWeather.Hardware;
 
-namespace WifiWeather
+namespace WifiWeather;
+
+internal class MainController
 {
-    internal class MainController
+    bool firstWeatherForecast = true;
+
+    private int currentGraphType = 0;
+
+    private IWifiWeatherHardware hardware;
+    private INetworkAdapter network;
+    private DisplayController displayController;
+    private RestClientController restClientController;
+
+    private List<double> temperatureReadings = new List<double>();
+    private List<double> pressureReadings = new List<double>();
+    private List<double> humidityReadings = new List<double>();
+
+    public MainController(IWifiWeatherHardware hardware, INetworkAdapter network)
     {
-        bool firstWeatherForecast = true;
+        this.hardware = hardware;
+        this.network = network;
+    }
 
-        private int currentGraphType = 0;
+    public void Initialize()
+    {
+        hardware.Initialize();
 
-        private IWifiWeatherHardware hardware;
-        private INetworkAdapter network;
-        private DisplayController displayController;
-        private RestClientController restClientController;
-
-        private List<double> temperatureReadings = new List<double>();
-        private List<double> pressureReadings = new List<double>();
-        private List<double> humidityReadings = new List<double>();
-
-        public MainController(IWifiWeatherHardware hardware, INetworkAdapter network)
+        hardware.UpButton.Clicked += (s, e) =>
         {
-            this.hardware = hardware;
-            this.network = network;
+            currentGraphType = currentGraphType - 1 < 0 ? 2 : currentGraphType - 1;
+
+            UpdateGraph();
+        };
+
+        hardware.DownButton.Clicked += (s, e) =>
+        {
+            currentGraphType = currentGraphType + 1 > 2 ? 0 : currentGraphType + 1;
+
+            UpdateGraph();
+        };
+
+        displayController = new DisplayController(hardware.Display);
+        restClientController = new RestClientController();
+
+        displayController.ShowSplashScreen();
+        Thread.Sleep(3000);
+        displayController.ShowDataScreen();
+    }
+
+    private void UpdateGraph()
+    {
+        switch (currentGraphType)
+        {
+            case 0:
+                displayController.UpdateGraph(currentGraphType, temperatureReadings);
+                break;
+            case 1:
+                displayController.UpdateGraph(currentGraphType, pressureReadings);
+                break;
+            case 2:
+                displayController.UpdateGraph(currentGraphType, humidityReadings);
+                break;
         }
+    }
 
-        public void Initialize()
+    async Task UpdateOutdoorValues()
+    {
+        displayController.UpdateSyncStatus(true);
+
+        var outdoorConditions = await restClientController.GetWeatherForecast();
+
+        if (outdoorConditions != null)
         {
-            hardware.Initialize();
+            firstWeatherForecast = false;
 
-            hardware.UpButton.Clicked += (s, e) =>
+            temperatureReadings.Add(outdoorConditions.Value.Item2);
+            pressureReadings.Add(outdoorConditions.Value.Item3);
+            humidityReadings.Add(outdoorConditions.Value.Item4);
+
+
+            if (temperatureReadings.Count > 10)
             {
-                currentGraphType = currentGraphType - 1 < 0 ? 2 : currentGraphType - 1;
-
-                UpdateGraph();
-            };
-
-            hardware.DownButton.Clicked += (s, e) =>
-            {
-                currentGraphType = currentGraphType + 1 > 2 ? 0 : currentGraphType + 1;
-
-                UpdateGraph();
-            };
-
-            displayController = new DisplayController(hardware.Display);
-            restClientController = new RestClientController();
-
-            displayController.ShowSplashScreen();
-            Thread.Sleep(3000);
-            displayController.ShowDataScreen();
-        }
-
-        private void UpdateGraph()
-        {
-            switch (currentGraphType)
-            {
-                case 0:
-                    displayController.UpdateGraph(currentGraphType, temperatureReadings);
-                    break;
-                case 1:
-                    displayController.UpdateGraph(currentGraphType, pressureReadings);
-                    break;
-                case 2:
-                    displayController.UpdateGraph(currentGraphType, humidityReadings);
-                    break;
-            }
-        }
-
-        async Task UpdateOutdoorValues()
-        {
-            displayController.UpdateSyncStatus(true);
-
-            var outdoorConditions = await restClientController.GetWeatherForecast();
-
-            if (outdoorConditions != null)
-            {
-                firstWeatherForecast = false;
-
-                temperatureReadings.Add(outdoorConditions.Value.Item2);
-                pressureReadings.Add(outdoorConditions.Value.Item3);
-                humidityReadings.Add(outdoorConditions.Value.Item4);
-
-
-                if (temperatureReadings.Count > 10)
-                {
-                    temperatureReadings.RemoveAt(0);
-                    pressureReadings.RemoveAt(0);
-                    humidityReadings.RemoveAt(0);
-                }
-
-                displayController.UpdateReadings(
-                    readingType: currentGraphType,
-                    icon: outdoorConditions.Value.Item1,
-                    temperature: outdoorConditions.Value.Item2,
-                    pressure: outdoorConditions.Value.Item3,
-                    humidity: outdoorConditions.Value.Item4,
-                    feelsLike: outdoorConditions.Value.Item5,
-                    sunrise: outdoorConditions.Value.Item6,
-                    sunset: outdoorConditions.Value.Item7);
-
-                UpdateGraph();
+                temperatureReadings.RemoveAt(0);
+                pressureReadings.RemoveAt(0);
+                humidityReadings.RemoveAt(0);
             }
 
-            displayController.UpdateSyncStatus(false);
+            displayController.UpdateReadings(
+                readingType: currentGraphType,
+                icon: outdoorConditions.Value.Item1,
+                temperature: outdoorConditions.Value.Item2,
+                pressure: outdoorConditions.Value.Item3,
+                humidity: outdoorConditions.Value.Item4,
+                feelsLike: outdoorConditions.Value.Item5,
+                sunrise: outdoorConditions.Value.Item6,
+                sunset: outdoorConditions.Value.Item7);
+
+            UpdateGraph();
         }
 
-        public async Task Run()
+        displayController.UpdateSyncStatus(false);
+    }
+
+    public async Task Run()
+    {
+        while (true)
         {
-            while (true)
+            displayController.UpdateWiFiStatus(network.IsConnected);
+
+            if (network.IsConnected)
             {
-                displayController.UpdateWiFiStatus(network.IsConnected);
+                int TimeZoneOffSet = -8; // PST
+                var today = DateTime.Now.AddHours(TimeZoneOffSet);
 
-                if (network.IsConnected)
+                Resolver.Log.Trace("Connected!");
+
+                if (today.Minute == 0 ||
+                    today.Minute == 10 ||
+                    today.Minute == 20 ||
+                    today.Minute == 30 ||
+                    today.Minute == 40 ||
+                    today.Minute == 50 ||
+                    firstWeatherForecast)
                 {
-                    int TimeZoneOffSet = -8; // PST
-                    var today = DateTime.Now.AddHours(TimeZoneOffSet);
+                    Resolver.Log.Trace("Getting forecast values...");
 
-                    Resolver.Log.Trace("Connected!");
+                    await UpdateOutdoorValues();
 
-                    if (today.Minute == 0 ||
-                        today.Minute == 10 ||
-                        today.Minute == 20 ||
-                        today.Minute == 30 ||
-                        today.Minute == 40 ||
-                        today.Minute == 50 ||
-                        firstWeatherForecast)
-                    {
-                        Resolver.Log.Trace("Getting forecast values...");
-
-                        await UpdateOutdoorValues();
-
-                        Resolver.Log.Trace("Forecast acquired!");
-                    }
-
-                    displayController.UpdateStatus(today.ToString("hh:mm tt | dd/MM/yyyy"));
-
-                    await Task.Delay(TimeSpan.FromMinutes(1));
+                    Resolver.Log.Trace("Forecast acquired!");
                 }
-                else
-                {
-                    Resolver.Log.Trace("Not connected, checking in 10s");
 
-                    await Task.Delay(TimeSpan.FromSeconds(10));
-                }
+                displayController.UpdateStatus(today.ToString("hh:mm tt | dd/MM/yyyy"));
+
+                await Task.Delay(TimeSpan.FromMinutes(1));
+            }
+            else
+            {
+                Resolver.Log.Trace("Not connected, checking in 10s");
+
+                await Task.Delay(TimeSpan.FromSeconds(10));
             }
         }
     }
