@@ -1,15 +1,14 @@
-﻿namespace WifiWeather
-{
-    public class Secrets
-    {
-        /// <summary>
-        /// Open Weather Map API key.
-        /// </summary>
-        public const string WEATHER_API_KEY = "WEATHER_API_KEY";
+﻿namespace WifiWeather;
 
-        /// <summary>
-        /// City to get weather for.
-        /// </summary>
-        public const string WEATHER_CITY = "WEATHER_CITY";
-    }
+public class Secrets
+{
+    /// <summary>
+    /// Open Weather Map API key.
+    /// </summary>
+    public const string WEATHER_API_KEY = "WEATHER_API_KEY";
+
+    /// <summary>
+    /// City to get weather for.
+    /// </summary>
+    public const string WEATHER_CITY = "WEATHER_CITY";
 }
